app/databuk/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zarr_fuse"
version = "0.2.0"
description = "Collecting and merging of scientific datasets from measurements and simulation sampling, further usable for creating various machine learning datasets."
readme = "README.md"
urls = { Homepage = "https://github.com/geomop/zarr_fuse" }
authors = [
  { name = "Jan Brezina", email = "jan.brezina@tul.cz" }
]
license = { text = "BSD-3-Clause" }
keywords = ["science", "datasets", "machine learning", "zarr", "data merging"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3"
]
requires-python = ">=3.11" # at least due to zarr>=3.x
dependencies = [
    "PyYAML",
    "attrs",
    "tzdata",
    "numpy",
    "scipy",    # implicit dependency of xarray interpolation
    "polars",
    "pint",
    "pint-xarray",
    "anytree",
    "pandas",
    "pyarrow",
    # "filterpy",  # commented out dependency
    "seaborn",
    # "pftools[all]",  # commented out dependency
    "zarr",       # 3.x does not support write_empty_chunks
    "dask",       # for compatibility with xarray 0.18.2
    "xarray",     # 0.19.0 does not support zarr 2.x well
    "s3fs",     # boto3 restriction => s3fs==0.4.2
    "fsspec",   # boto3 restriction => fsspec==2025.5.1
    "flask",
    "python-dotenv"
]

[project.optional-dependencies]
plot=[
    "matplotlib",
    "plotly",
    "contextily",
    "pyproj"
]
apps=[
    "joblib",
    "requests",
    "requests_cache",
    "ipykernel",
    "jupyter",
    #cdsapi      # era5 weather data
    #cfgrib      # requirement for xarray GRIB file support
]
dev = [
    # Have to repeat 'apps' and 'plotting' here
    # TODO: find a way to avoid this repetition
    # possibly PDM (Python Development Master or stick with basic setup.py
   "pytest",
    # Include dependencies from plotting
    "matplotlib",
    "plotly",
    # Include dependencies from apps
    "joblib",
    "requests",
    "requests_cache",
    "ipykernel",
    "jupyter"
    # any additional development tools
]

[tool.setuptools.packages.find]
where = ["."]
include = ["zarr_fuse", "zarr_fuse.*"]
